+++ src/onegov.form/onegov/form/tests/test_fields.py
@@ -10,6 +10,8 @@ from onegov.form.fields import HtmlField
 from onegov.form.fields import MultiCheckboxField
 from onegov.form.fields import OrderedMultiCheckboxField
 from onegov.form.fields import UploadField
+from onegov.form.fields import PhoneNumberField
+from onegov.form.validators import ValidPhoneNumber
 
 
 def create_file(mimetype, filename, content):
@@ -82,3 +84,34 @@ def test_html_field():
     field.data = '<script>alert(0)</script>'
     field.validate(form)
     assert '<script>' not in field.data
+
+
+def test_phone_number_field():
+    form = Form()
+    field = PhoneNumberField()
+    field = field.bind(form, 'phone_number')
+    assert field.country == 'CH'
+    assert field.validators[0].country == 'CH'
+
+    field.data = ''
+    assert field.formatted_data == ''
+    assert field.validate(form)
+
+    field.data = '791112233'
+    assert field.formatted_data == '+41791112233'
+    assert field.validate(form)
+
+    field.data = 'abc'
+    assert field.formatted_data == 'abc'
+    assert not field.validate(form)
+
+    form = Form()
+    field = PhoneNumberField(country='DE')
+    field = field.bind(form, 'phone_number')
+    assert field.validators[0].country == 'DE'
+    assert field.country == 'DE'
+
+    form = Form()
+    field = PhoneNumberField(validators=[ValidPhoneNumber(country='DE')])
+    field = field.bind(form, 'phone_number')
+    assert field.validators[0].country == 'DE'